Computes covariances of composite scores given a covariance matrix and a weight matrix

Computes covariances of composite scores given a covariance matrix and a weight matrix

Usage

composite_covariance(Sigma, w, correlation = FALSE)

Arguments

Sigma

Covariance matrix

w

Weight matrix. Must have the same number of rows as R

correlation

If TRUE, return correlations instead of covariances

Examples

# Create covariance matrix
Sigma <- matrix(0.6, nrow = 5, ncol = 5)
diag(Sigma) <- 1
# Create weight matrix
w <- matrix(0, nrow = 5, ncol = 2)
w[1:2,1] <- 1
w[3:5,2] <- 1
w
#>      [,1] [,2]
#> [1,]    1    0
#> [2,]    1    0
#> [3,]    0    1
#> [4,]    0    1
#> [5,]    0    1
# covariance matrix of weighted sums
composite_covariance(Sigma, w)
#>      [,1] [,2]
#> [1,]  3.2  3.6
#> [2,]  3.6  6.6